fix: refactor class imports
